♻ refactor: Integer로 키 타입 변경 (long은 프론트에서 커버 불가)

+++ OneCoin/Server/src/test/java/OneCoin/Server/helper/StubData.java
@@ -20,7 +20,7 @@ public class StubData {
         public static Order getMockEntity() {
             return Order
                     .builder()
-                    .orderId(1L)
+                    .orderId(1)
                     .limit(new BigDecimal("22525000"))
                     .amount(new BigDecimal("10"))
                     .completedAmount(BigDecimal.ZERO)
@@ -30,7 +30,7 @@ public class StubData {
                     .build();
         }
 
-        public static Order getMockEntity(long orderId, String limit,
+        public static Order getMockEntity(int orderId, String limit,
                                           String amount, String orderType, String code, long userId) {
             return Order
                     .builder()
@@ -45,11 +45,11 @@ public class StubData {
         }
 
         public static List<Order> getMockEntities() {
-            Order mockEntity1 = MockOrder.getMockEntity(2L, "333333", "100", "BID", "KRW-BTC", 2L);
-            Order mockEntity2 = MockOrder.getMockEntity(3L, "333333", "200", "ASK", "KRW-ETH", 3L);
-            Order mockEntity3 = MockOrder.getMockEntity(4L, "555555", "300", "ASK", "KRW-ETH", 4L);
-            Order mockEntity4 = MockOrder.getMockEntity(5L, "333333", "400", "ASK", "KRW-ETH", 5L);
-            Order mockEntity5 = MockOrder.getMockEntity(6L, "555555", "500", "BID", "KRW-XRP", 6L);
+            Order mockEntity1 = MockOrder.getMockEntity(2, "333333", "100", "BID", "KRW-BTC", 2L);
+            Order mockEntity2 = MockOrder.getMockEntity(3, "333333", "200", "ASK", "KRW-ETH", 3L);
+            Order mockEntity3 = MockOrder.getMockEntity(4, "555555", "300", "ASK", "KRW-ETH", 4L);
+            Order mockEntity4 = MockOrder.getMockEntity(5, "333333", "400", "ASK", "KRW-ETH", 5L);
+            Order mockEntity5 = MockOrder.getMockEntity(6, "555555", "500", "BID", "KRW-XRP", 6L);
 
             return List.of(mockEntity1, mockEntity2, mockEntity3, mockEntity4, mockEntity5);
         }
